refactor of files to folder
